# Conflicts:
#	factory/core/coreComponents.go

+++ node/metrics/metrics_test.go
@@ -309,7 +309,7 @@ func TestInitRatingsMetrics(t *testing.T) {
 
 	ash := &statusHandler.AppStatusHandlerStub{
 		SetUInt64ValueHandler: func(key string, value uint64) {
-			keys[key] = uint64(value)
+			keys[key] = value
 		},
 		SetStringValueHandler: func(key string, value string) {
 			keys[key] = value
